extacted messages to messages file
